Tidy up API some more
