+++ packages/ciphernode/enclave_node/src/ciphernode.rs
@@ -3,12 +3,12 @@ use alloy::primitives::Address;
 use anyhow::Result;
 use data::Data;
 use enclave_core::EventBus;
-use evm::{CiphernodeRegistrySol, CiphernodeRegistrySolReader, EnclaveSolReader};
+use evm::{CiphernodeRegistrySol, EnclaveSolReader};
 use logger::SimpleLogger;
 use p2p::P2p;
 use rand::SeedableRng;
 use rand_chacha::rand_core::OsRng;
-use router::{CiphernodeSelector, CommitteeMetaFactory, E3RequestRouter, LazyFhe, LazyKeyshare};
+use router::{CiphernodeSelector, E3RequestRouter, LazyFhe, LazyKeyshare};
 use sortition::Sortition;
 use std::sync::{Arc, Mutex};
 use tokio::task::JoinHandle;
