+++ packages/ciphernode/evm/src/lib.rs
@@ -2,7 +2,11 @@ mod enclave_sol;
 mod ciphernode_registry_sol;
 mod registry_filter_sol;
 pub mod helpers;
+mod enclave_sol_reader;
+mod enclave_sol_writer;
 
 pub use enclave_sol::*;
+pub use enclave_sol_reader::*;
+pub use enclave_sol_writer::*;
 pub use ciphernode_registry_sol::*;
 pub use registry_filter_sol::*;
